♻️ :: 권한 허용 위치 변경
