+++ presentation/src/main/java/com/semicolon/walkhub/ui/login/LoginActivity.kt
@@ -48,8 +48,6 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(
         repeatOnStarted {
             vm.eventFlow.collect { event -> handleEvent(event) }
         }
-
-        requestPermissions()
     }
 
     private fun handleEvent(event: LoginViewModel.Event) = when (event) {
@@ -83,52 +81,6 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(
         else if (id.isNotEmpty() && password.isNotEmpty()){
 
         }
-    }
-
-    private val fitnessOptions =
-        FitnessOptions.builder()
-            .addDataType(DataType.AGGREGATE_STEP_COUNT_DELTA)
-            .addDataType(DataType.AGGREGATE_MOVE_MINUTES)
-            .addDataType(DataType.AGGREGATE_DISTANCE_DELTA)
-            .addDataType(DataType.AGGREGATE_CALORIES_EXPENDED)
-            .addDataType(DataType.TYPE_LOCATION_SAMPLE)
-            .build()
-
-    private fun requestPermissions() {
-        val permissionListener = object : PermissionListener {
-            override fun onPermissionGranted() {
-                fitSignIn()
-            }
-
-            override fun onPermissionDenied(deniedPermissions: List<String>) {
-                finish()
-            }
-        }
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            TedPermission.create()
-                .setPermissionListener(permissionListener)
-                .setPermissions(
-                    Manifest.permission.ACTIVITY_RECOGNITION,
-                    Manifest.permission.ACCESS_FINE_LOCATION,
-                    Manifest.permission.ACCESS_COARSE_LOCATION
-                ).check()
-        } else {
-            TedPermission.create()
-                .setPermissionListener(permissionListener)
-                .setPermissions(
-                    Manifest.permission.ACCESS_FINE_LOCATION,
-                    Manifest.permission.ACCESS_COARSE_LOCATION
-                ).check()
-        }
-    }
-
-    private fun fitSignIn() {
-        GoogleSignIn.requestPermissions(
-            this, 0,
-            GoogleSignIn.getAccountForExtension(this, fitnessOptions),
-            fitnessOptions
-        )
     }
 
     override fun initView() {
